+++ src/modules/Watchlist/services/watchlist-storage.service.ts
@@ -1,10 +1,11 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams, HttpHeaders } from '@angular/common/http';
-import { Observable } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { map } from "rxjs/operators";
 import { IWatchlistItem } from './watchlist.service';
 import { AppConfigService } from '@app/services/app.config.service';
 import { IInstrument } from '@app/models/common/instrument';
+import { IdentityService } from '@app/services/auth/identity.service';
 
 @Injectable()
 export class WatchlistStorageService {
@@ -13,9 +14,13 @@ export class WatchlistStorageService {
         return `${AppConfigService.config.apiUrls.userDataStoreREST}Watchlists`;        
     }
 
-    constructor(private http: HttpClient) { }
+    constructor(private http: HttpClient, private _identity: IdentityService) { }
 
     public allWatchlists(): Observable<IWatchlistItem[]> {
+        if (this._identity.isGuestMode) {
+            return of([]);
+        }
+        
         return this.http.get<IWatchlistItem[]>(`${this._templatesURL}/all`)
             .pipe(map((data: any) => {
                 return data.map((value: any) => {
@@ -29,6 +34,10 @@ export class WatchlistStorageService {
     }
 
     public saveWatchlist(name: string, data: IInstrument[]): Observable<IWatchlistItem> {
+        if (this._identity.isGuestMode) {
+            return of(null);
+        }
+
         let postData = {
             Name: name,
             Data: data,
@@ -44,6 +53,10 @@ export class WatchlistStorageService {
     }
 
     public editWatchlistName(id: string, name: string): Observable<any> {
+        if (this._identity.isGuestMode) {
+            return of(null);
+        }
+
         let patchData = {
             Id: id,
             PropertyName: "Name",
@@ -54,6 +67,10 @@ export class WatchlistStorageService {
     }
     
     public editWatchlist(id: string, newData: IInstrument[], name: string): Observable<any> {
+        if (this._identity.isGuestMode) {
+            return of(null);
+        }
+
         let patchData = {
             Id: id,
             Name: name,
@@ -64,6 +81,10 @@ export class WatchlistStorageService {
     }
 
     public removeWatchlist(id: string): Observable<any> {
+        if (this._identity.isGuestMode) {
+            return of(null);
+        }
+
         let params = new HttpParams();
         params = params.append("id", id);
         return this.http.delete(this._templatesURL, { params: params });
